% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ank78.indices}
\alias{ank78.indices}
\title{ank78.indices}
\format{
an \code{FLIndices} object
}
\usage{
ank78.indices
}
\description{
Survey indices for black-bellied anglerfish in the Celtic Sea and northern Bay of Biscay
}
\author{
Hans Gerritsen
}
