% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ank78}
\alias{ank78}
\title{ank78}
\format{
an \code{FLStock} object
}
\usage{
ank78
}
\description{
Catch numbers, catch weights, etc. for black-bellied anglerfish stock in the Celtic Sea and oorthern Bay of Biscay
}
\examples{

data(ank78)
summary(ank78)
}
\author{
Hans Gerritsen
}
