% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{FLR_fakehake}
\alias{FLR_fakehake}
\title{Stock with similar life histroy traits to hake created with FLife}
\format{
A FLStock object
\describe{
\item{slot}{these are the slots form FLR, e.g. catch.n, catch.wt, m, stock.n, stock....etc.}
\item{age}{}
\item{area}{}
\item{data}{}
\item{iter}{}
\item{season}{}
\item{unit}{}
\item{year}{}
...
}
}
\usage{
FLR_fakehake
}
\description{
A FLStock object
}
\keyword{datasets}
