% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.sbarclass.R
\name{plot.sbarclass}
\alias{plot.sbarclass}
\title{Plot object of with class 'sbarclass'}
\usage{
\method{plot}{sbarclass}(x, out = "fit", ...)
}
\arguments{
\item{x}{object of class:sbarclass}

\item{out}{a character that specifies whether the "fit" or "stock" should be plotted}

\item{...}{No specific usage}
}
\value{
list
}
\description{
Plot an object with class:'sbarclass', plot fit or stock predictions with ggplot functions
}
\details{
See vignette for a detailed example of this function in use.
}
\examples{
\dontrun{ 
plot(x,out="fit")
  plot(x,out="stock")}
}
\seealso{
\code{\link{makesbarclass}} for to conversion into an object that can be used by this function.
}
