% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schnute_obserror.R
\name{schnute_obserror}
\alias{schnute_obserror}
\title{Prepare list for Schnute adapted observation error assessment for an optimiser}
\usage{
schnute_obserror(
  version = 2,
  catch_b,
  indices_b,
  ts,
  mwts,
  tsp = 0,
  rho,
  W,
  ind_l_wt = 1,
  start_q = 1e-08,
  start_indexsigma = 0.1,
  start_B0,
  start_sigma = exp(-0.2),
  start_f_calc = 0.3,
  start_rec_a,
  start_rec_b,
  spawn_prop = 1,
  start_catchsigma = 0.1,
  fix_sigma = TRUE,
  fix_B0 = FALSE,
  fix_indexsigma = FALSE,
  fix_catchsigma = TRUE,
  adrep = FALSE
)
}
\arguments{
\item{version}{numeric, either 1, 2 or 3. This controls what deterministic equations in the model are used to derive population biomass. 1 and 2 use the fraction of of total biomass in a given year due to newly recruited fish. This fraction is derived from mean weights and detailed in the \strong{schnute} vignette. \code{version = 3} is the more classical population dynamics.
\tabular{cc}{
1 \tab whole biomass derived from recruit biomass \cr
2 \tab whole biomass derived from previously exploited biomass \cr
3 \tab whole biomass is a combination of recruit biomass and previously exploited biomass}}

\item{catch_b}{numeric vector of catch biomass over time period of assessment}

\item{indices_b}{matrix of biomass surveys (CPUE) of dimensions: no. of surveys x no.years}

\item{ts}{numeric. Survey timing parameters}

\item{mwts}{matrix of mean weights from sampling with dimensions: 3 x no. years. recruit mean weights \eqn{\bar{Y}} (first row), previously exploited biomass mean weights \eqn{\bar{Z}} (second row) and entire assessed biomass mean weight \eqn{\bar{X}} (third row).}

\item{tsp}{numeric. Timing of spawning. Default to 0 (start of year).}

\item{rho}{numeric. Growth parameter, slope of linear growth model.}

\item{W}{numeric. Growth parameter, intercept of linear growth model.}

\item{ind_l_wt}{numeric. Survey weighting in the likelihood. Defaults to 1 fro each survey, ie.e. equal weighting}

\item{start_q}{Starting values for survey catchability parameters. Default is 1e-6}

\item{start_indexsigma}{Starting values for survey sd parameters. Default is 0.1}

\item{start_B0}{Starting parameter value for biomass at first time step. Default is 5*max(catch_b)}

\item{start_sigma}{Starting parameter value fraction of population that survives natural moratlity. Default is \eqn{e^{0.2}}}

\item{start_f_calc}{Starting parameter values for estimated fishing mortality. Default is 0.3.}

\item{start_rec_a}{Starting parameter value for the 'a' parameter of the Beverton-Holt stock-recruit function. The asymptotic biomass of recruits. Default is 1/5*max(catch_b).}

\item{start_rec_b}{Starting parameter value for the 'b' parameter of the Beverton-Holt stock-recruit function. The spawning stock biomass needed to produce a/2 on average. Default is 4*max(catch_b).}

\item{spawn_prop}{proportion of biomass that is mature. Defaults to 1 for each year.}

\item{start_catchsigma}{Starting parameter value for catch sd. Default is 0.1}

\item{fix_sigma}{logical}

\item{fix_B0}{logical}

\item{fix_indexsigma}{logical}

\item{fix_catchsigma}{logical}

\item{adrep}{logical. Whether the user would like the ADreport variables (and their derivatives) reported for starting parameters.}
}
\value{
List with components for optimiser in R. This output is that of the function \link[TMB]{MakeADFun} from TMB
}
\description{
Create list with TMB framework, including data, gradients and NLL function for a Schnute adapted observation error assessment that can be optimised.
}
