% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schnute.R
\name{schnute_peb}
\alias{schnute_peb}
\title{Prepare list for SChnute assessment in an optimiser}
\usage{
schnute_peb(
  version,
  catch,
  indices,
  ts,
  mwts,
  tsp = 0,
  rho,
  W,
  start_q = 1e-08,
  start_indexsigma = 0.1,
  start_B0 = sb0,
  start_sigma = exp(-0.2),
  start_f_calc = 0.3,
  start_catchsigma = 0.1,
  fix_sigma = TRUE,
  fix_B0 = FALSE,
  fix_indexsigma = FALSE,
  fix_catchsigma = TRUE
)
}
\arguments{
\item{version}{character}

\item{catch}{numeric}

\item{indices}{matrix}

\item{ts}{numeric}

\item{mwts}{matrix}

\item{tsp}{numeric}

\item{rho}{numeric}

\item{W}{numeric}

\item{start_q}{numeric}

\item{start_indexsigma}{numeric}

\item{start_B0}{numeric}

\item{start_sigma}{numeric}

\item{start_f_calc}{numeric}

\item{start_catchsigma}{numeric}

\item{fix_sigma}{logical}

\item{fix_B0}{logical}

\item{fix_indexsigma}{logical}

\item{fix_catchsigma}{logical}
}
\value{
list
}
\description{
Create an object with TMB framework, including data, gradients and NLL function that can be optimised
}
