% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbarclass.R
\name{makesbarclass}
\alias{makesbarclass}
\title{Convert sdreport summary into an object of for plotting with class: sbarclass
#'
Create an object with class:'sbarclass', which can then be used to quickly plot fit and stock predictions from an sbar assessment}
\usage{
makesbarclass(x = matrix(), survey_names, cat, ind, years)
}
\arguments{
\item{x}{matrix}

\item{survey_names}{character}

\item{cat}{numeric}

\item{ind}{matrix}

\item{years}{numeric}
}
\value{
list
}
\description{
Convert sdreport summary into an object of for plotting with class: sbarclass
#'
Create an object with class:'sbarclass', which can then be used to quickly plot fit and stock predictions from an sbar assessment
}
