% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.sbarclass.R
\name{plot.sbarclass}
\alias{plot.sbarclass}
\title{Plot object of with class 'sbarclass'}
\usage{
\method{plot}{sbarclass}(x, out = "fit", ...)
}
\arguments{
\item{x}{sbar}

\item{out}{character}

\item{...}{No specific usage}
}
\value{
list
}
\description{
Plot object with class:'sbarclass', plot fit or stock predictions
}
