% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbar-package.R
\docType{package}
\name{sbar-package}
\alias{sbar}
\alias{sbar-package}
\title{sbar: Stage-based assessments in R}
\description{
Implementations of the stage-based fisheries assessment models CSA and Schnute (1987). Models are those detailed and implemented in a manuscript currently under review.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/lbatts/sbar}
  \item Report bugs at \url{https://github.com/lbatts/sbar/issues}
}

}
\author{
\strong{Maintainer}: Luke Batts \email{luke.batts@research.gmit.ie}

Authors:
\itemize{
  \item Coilin Minto \email{coilin.minto@gmit.ie}
}

Other contributors:
\itemize{
  \item Hans Gerritsen [contributor]
}

}
\keyword{internal}
