% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schnute_orig.R
\name{schnute_orig}
\alias{schnute_orig}
\title{Prepare list for SChnute assessment in an optimiser}
\usage{
schnute_orig(
  version = 2,
  catchkg,
  indiceskg,
  ts,
  mwts,
  tsp = 0,
  mu = 0.5,
  rho,
  W,
  ind_l_wt = 1,
  start_q = 1e-08,
  start_indexsigma = 0.1,
  start_sigma = exp(-0.2),
  start_rec_a,
  start_rec_b,
  fix_sigma = TRUE,
  fix_indexsigma = FALSE
)
}
\arguments{
\item{version}{numeric}

\item{catchkg}{numeric}

\item{indiceskg}{matrix}

\item{ts}{numeric}

\item{mwts}{matrix}

\item{tsp}{numeric}

\item{mu}{numeric}

\item{rho}{numeric}

\item{W}{numeric}

\item{ind_l_wt}{numeric}

\item{start_q}{numeric}

\item{start_indexsigma}{numeric}

\item{start_sigma}{numeric}

\item{start_rec_a}{numeric}

\item{start_rec_b}{numeric}

\item{fix_sigma}{logical}

\item{fix_indexsigma}{logical}
}
\value{
list
}
\description{
Create an object with TMB framework, including data, gradients and NLL function that can be optimised
}
